Rename namespace and file
